ADM-350 update loading component and test
