+++ frontend/__tests__/src/App.test.tsx
@@ -1,10 +1,10 @@
-import { render } from '@testing-library/react'
+import { render, RenderResult } from '@testing-library/react'
 import App from '@src/App'
 import { Provider } from 'react-redux'
 import { store } from '@src/store'
 
 describe('render app', () => {
-  const setup = () => {
+  const setup = (): RenderResult => {
     return render(
       <Provider store={store}>
         <App />
@@ -12,15 +12,13 @@ describe('render app', () => {
     )
   }
   it('should show hello World when render app', () => {
-    const { getByText, rerender } = setup()
+    const { container, rerender } = setup()
 
     rerender(
       <Provider store={store}>
         <App />
       </Provider>
     )
-    const loadText = getByText('Loading...')
-
-    expect(loadText).toBeInTheDocument()
+    expect(container.getElementsByTagName('span')[0].getAttribute('role')).toEqual('progressbar')
   })
 })
