+++ frontend/src/components/Loading/index.tsx
@@ -1,0 +1,10 @@
+import { CircularProgress } from '@mui/material'
+import { LoadingDrop } from './style'
+
+export const Loading = () => {
+  return (
+    <LoadingDrop open>
+      <CircularProgress size='8rem' />
+    </LoadingDrop>
+  )
+}
